fix: activate current match in subsequent calls to update_search
This activates the currently activated match position in subsequent calls
to update_search made by methods that don't set activate_match_pos
attribute after the CopyMode KeyAssignment SearchForwardRelativeToCursor
is triggered and the first match position after the cursor is activated.
It is only required to set activate_match_pos attribute to AtCursor
if activate_match_pos attribute is AfterCursor because the currently
activated match position is activated if the attribute is BeforeCursor.