feat: add cyclic switches to EditCommand struct
Cyclic switches allow to switch between multiple choices for a flag and
even toggle the flag. They are similar to options in the sense that they
can accept one of many choices and be toggled on and off. They are
similar to switches in the sense that pressing the specific key
combination just updates the value for flag without requiring any input
from user.
It accepts a list of strings named choices, and an optional default
value. If default value is not specified or is not found in the choices
list, then the flag is initially not set. Else, the flag is set to the
default value.
To switch between choices, user needs to press the keys specified for
the flag.
It accepts a boolean field called allow_nil, which if true resets flag
value when user enters specified keys and the flag was set to the last
choice at the time of entering the keys. If false, the flag is set to
the first choice. Hence why this struct is called CyclicSwitch