feat: add TabulatedList KeyAssignment
TabulatedList overlay displays a list of provided choices with an
ability to through the list up and down using the keys Ctrl-P and Ctrl-N
respectively.
TabulatedList accepts below arguments:
* description
* context (optional arguments that used to display additional
  information)
* choices (list of items to be displayed and moved through in the
  overlay)
* actions: list of TransientArgument objects that can be passed

Structure of TransientArgument is as follows:
```
{
    key = String,
    description = String,
    action = wezterm.action_callback(function(window, pane, line)
    end)
}
```

Each TransientArgument object indicates a key action pair. Entering the
specific key will trigger the action listed against action field

Context field is useful when composing with TransientMenu in that the
state set in the TransientMenu can be displayed under context in
TabulatedList overlay