feat: add key to copy active choice to clipboard in SelectorActions
Add 'y' key to copy active choice to clipboard in SelectorActions
KeyAssignment. If id is present for current choice, then id is copied to
clipboard else label is copied to clipboard. If 'y' belongs to entries
of current node, then the key is consumed and the specific node is
traversed into.