perf: truncate traversed_nodes to size 1 instead of assigning to vector
Truncate traversed_nodes to size 1 instead of assigning it to vector
containing root node because the first item in traversed_nodes vector is
the root node. This way we don't have to reassign traversed_nodes to a
new vector containing root_node. We also don't need to store a separate
reference to root_node because it is already present in travresed_nodes