refactor: group logic for rendering changes and polling input
Move all logic to render changes to the render method in TransientState
struct. This is in order to make it easier to track changes.
Compute changes to redraw entire screen because only the minimum set of
changes is applied to the surface.
Move logic to poll input to the run_loop method in TransientState struct
and individually check if prompt or selector is active based on the
value of self.mode. This is in order to make it easier to render the
changes after handling the input